feat: refactoring NaviButtons abgeschlossen
