+++ src/components/buttons/navi-buttons/settings-button.tsx
@@ -1,8 +1,11 @@
 import React, { FC } from 'react';
 import { NaviButton } from './navi-button';
-import { IconSettings } from '../icons/icon-settings';
+import { IconSettings } from '../../icons/icon-settings';
 
 export type SettingsButtonProps = {
+  /**
+   * Specifies the action, which is called as the user clicks on the settings button.
+   */
   onClick: () => void;
 };
