+++ src/components/buttons/public-api.ts
@@ -1,9 +1,9 @@
-export * from './text-button';
 export * from './icon-button';
+export * from './interaction-buttons/like';
 export * from './interaction-buttons/reply';
 export * from './interaction-buttons/share';
-export * from './interaction-buttons/like';
-export * from './navi-button';
-export * from './logout-button';
-export * from './settings-button';
-export * from './profile-picture-button';
+export * from './navi-buttons/logout-button';
+export * from './navi-buttons/navi-button';
+export * from './navi-buttons/profile-picture-button';
+export * from './navi-buttons/settings-button';
+export * from './text-button';
