+++ src/layout/page-header/page-header.stories.tsx
@@ -3,10 +3,10 @@ import { action } from '@storybook/addon-actions';
 import React, { useState } from 'react';
 import { PageHeader } from './page-header';
 import { MumbleWhiteHorizontal } from '../../components/logos/mumble-white-horizontal';
-import { ProfilePictureButton } from '../../components/buttons/profile-picture-button';
+import { ProfilePictureButton } from '../../components/buttons/navi-buttons/profile-picture-button';
 import { Navigation } from '../navigation/navigation';
-import { SettingsButton } from '../../components/buttons/settings-button';
-import { LogoutButton } from '../../components/buttons/logout-button';
+import { SettingsButton } from '../../components/buttons/navi-buttons/settings-button';
+import { LogoutButton } from '../../components/buttons/navi-buttons/logout-button';
 import { Modal, ModalType } from '../../components/modal/modal';
 import { Label, LabelSizes } from '../../components/typography/label';
 import { Form } from '../../components/form/form';
